requirements.txt
fastapi==0.79.1
strawberry-graphql==0.126.0
uvicorn==0.18.2
